Extract zone_import_export_test
